[CI] Split Kimi-K3 B300 low-latency coverage
